chronos/requirements.txt
requests==2.18.4
mysqlclient==1.3.12
MySQL-python==1.2.5
python-crontab==2.2.8